separate ProjectileStats from Utility.cs
