+++ Assets/Scripts/Stats/ProjectileStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b56d40f47df29647b11dba188c597f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
